Add util method for strict Retrofit sync execution
Since HTTP response errors were being handled in the same way as I/O
errors, namely by showing an error message to the user, they were
usually handled through a custom dedicated exception class since the
upgrade to Retrofit 2 in commit
29f51f5b39757b54582b2bddc2096380663e8938. This was implemented in a
generic way for the asynchronous API via the base callback classes.

While this covered the overwhelming majority of the usages, but there
was no similar generic utility for the synchronous API, which would be
used more frequently after the courseware calls were converted to use
Retrofit (since they query the cache in many places).

This commit addresses this lack by adding an executeStrict() utility
method in a new CallUtil class for the synchronous API, which throws the
exception upon receiving an error status from the server, and returns
the converted body otherwise. This utility method is imported statically
for brevity.