Fix OkHttp callbacks not delivered on main thread
OkHttp delivers invokes it’s callbacks on the same worker thread on
which the request is queried, which Retrofit then delivers on the main
thread in the Android platform. This was causing issues on the Handouts
and Announcements queries, which were performed directly through OkHttp
since their URLs were resolved dynamically, but their callbacks were
assumed to have been delivered on the main thread. This commit fixes the
issue by modifying our base OkHttp callback implementations (OkCallback
and ErrorHandlingOkCallback) to deliver the result to their own abstract
callback methods on the main thread.