Fix custom cache query interceptor response
The interceptor previously delivered the original response object as the
inner cache response field in the returned response, which crashed
because it wasn’t supposed to have a network response assosiated with
it. This commit fixes the issue by explicitly removing the network,
prior, and cache response fields from the cache response field.

A constant TLS handshake is also added to all the response objects for
HTTPS requests, since this is required for validation by the
CacheStrategy resolver.

The CacheManager has also been modified to not log FileNotFoundException
errors encountered from the get() method like other IOException errors,
since the previous explicit check beforehand had been removed to make it
atomic.