Fix courseware cache not being transferred
After the Retrofit migration, the courseware cache was no longer
accessible from the CacheManager, due to the fact that the previous
implementation stripped the query parameters from the URL before storing
it in the cache. The course ID was also not URL-escaped. This commit
fixes the issue by adding logic in the cache retriever to check the URL,
and strip the parameters (and decode the course ID) upon encountering
the relevant one.