Don’t show loading errors from NavigationFragment
NavigationFragment is added by default in BaseSingleFragmentActivity,
but blocked from opening in all but the course listing and videos
screens. However, it’s still initialized, and loads the account data. It
also attempts to show an error message on whatever Activity it’s added
on if there an issue with the data loading.

This commit removes the logic in NavigationFragment for showing error
messages, since most of the time the drawer isn’t even being displayed
to the user. Even when it _is_ on display, the Activity is the wrong
place to show the error message on, since it overlays the Activity.
Moreover, since we already have the account data cached, there is not
reason to show an error message at all if we fail to load it from the
server, even if there had been a natural place to show such an error
message.